# public/swagger.json
{
	"swagger": "2.0",
	"info": {
		"version": "1.0.0",
		"title": "Open Banking Nigeria",
...
		"contact": {}
	},
	"host": "obsandbox.herokuapp.com",
	"basePath": "/",
	"securityDefinitions": {
		"Bearer": {
			"type": "apiKey",
			"name": "Authorization",
			"in": "header"
		}
	},
	"schemes": [
		"https"
	],
	"security": [
		{
			"Bearer": []
		}
	],
	"consumes": [
		"application/json"
	],
	"produces": [
		"application/json"
	],
	"paths": {
...
		"/v1/accounts/": {
			"get": {
				"description": "This endpoint retrieves all accounts belonging to a customer.\r\n\r\n`This endpoint supports paging`\r\n\r\n#### URL Parameters\r\n| Parameter |Type |Description |\r\n|--|--|--|\r\n| customerId|string |`Required` A unique identifier for the customer as defined by the bank|\r\n",
				"summary": "Get All Accounts",
				"tags": [
					"Accounts"
				],
				"operationId": "V1AccountCustomerByCustomerIdGet",
				"deprecated": false,
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "customerId",
						"in": "header",
						"required": true,
						"type": "string",
						"description": ""
					}
				],
				"responses": {
					"200": {
						"description": "",
						"schema": {
							"$ref": "#/definitions/GetAllAccounts%28SingleAccount%29"
						},
						"examples": {
							"application/json": {
								"status": "00",
								"message": "The process was completed successully",
								"data": {
									"page_info": {
										"total": 11,
										"current_page": 1,
										"total_pages": 2
									},
									"accounts": [
										{
											"accountNumber": "91000293",
											"customerId": 1,
											"accountName": "LAWAL ABU",
											"currency": "NGN",
											"accountOpeningDate": "2019-01-09T23:00:00.000Z",
											"accountType": "SAVINGS",
											"bvn": "293000",
											"fullName": "LAWAL ABU",
											"phoneNumber": 32920,
											"email": "labu@labu.ng",
											"status": "ACTIVE"
										},
										{
											"accountNumber": "9100029412",
											"customerId": 1,
											"accountName": "DEMOLA SANNI",
											"currency": "NGN",
											"accountOpeningDate": "2019-01-09T23:00:00.000Z",
											"accountType": "SAVINGS",
											"bvn": "293000",
											"fullName": "DEMOLA SANNI",
											"phoneNumber": 32920,
											"email": "DEMOLASANNI@labu.ng",
											"status": "ACTIVE"
										}
									]
								}
							}
						},
						"headers": {}
					}
				}
			}
		},
		"/v1/accounts/{accountNumber}": {
			"get": {
				"description": "This endpoint retrieves a specific account.\n\n\n| Parameter |Type |Description |\n|--|--|--|\n| accountNumber|string |`Required` Customer's account number in `NUBAN` format|\n|\n\n",
				"summary": "Get Specific Account",
				"tags": [
					"Accounts"
				],
				"operationId": "V1AccountByAccountNumberGet",
				"deprecated": false,
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "accountNumber",
						"in": "path",
						"required": true,
						"type": "string",
						"description": ""
					}
				],
				"responses": {
					"200": {
						"description": "",
						"schema": {
							"$ref": "#/definitions/GetSpecificAccount-Success"
						},
						"examples": {
							"application/json": {
								"status": "00",
								"message": "The process was completed successully",
								"data": [
									{
										"accountNumber": "1010119934",
										"customerId": "1234567890",
										"accountName": "Ado John Sule",
										"currency": "NGN",
										"accountOpeningDate": "2017-01-13",
										"lastTransactionDate": "2019-03-15T07:05:59.524Z",
										"accountType": "Wonderful Savings Account",
										"bvn": "221422456789",
										"fullName": "Ado John Sule",
										"phoneNumber": 342391,
										"email": "ado.john@example.com",
										"status": "Active",
										"expansionFields": [
											{
												"id": "Some random ID",
												"description": "Some random text",
												"type": "Some random type",
												"value": "Some value"
											}
										]
									}
								]
							}
						},
						"headers": {}
					}
				}
			}
		},
...
	},
	"definitions": {
...
		"ExpansionField": {
			"title": "ExpansionField",
			"example": {
				"id": "Some random ID",
				"description": "Some random text",
				"type": "Some random type",
				"value": "Some value"
			},
			"type": "object",
			"properties": {
				"id": {
					"type": "string"
				},
				"description": {
					"type": "string"
				},
				"type": {
					"type": "string"
				},
				"value": {
					"type": "string"
				}
			},
			"required": [
				"id",
				"description",
				"type",
				"value"
			]
		},
...
		"Data19": {
			"title": "GetAllAccounts(SingleAccount)",
			"example": {
				"accountNumber": "1010119934",
				"customerId": "1234567890",
				"accountName": "Ado John Sule",
				"currency": "NGN",
				"accountOpeningDate": "2017-01-13",
				"lastTransactionDate": "2019-03-15T07:05:59.524Z",
				"accountType": "Wonderful Savings Account",
				"bvn": "221422456789",
				"fullName": "Ado John Sule",
				"phoneNumber": 342391,
				"email": "ado.john@example.com",
				"status": "Active",
				"expansionFields": [
					{
						"id": "Some random ID",
						"description": "Some random text",
						"type": "Some random type",
						"value": "Some value"
					}
				]
			},
			"type": "object",
			"properties": {
				"accountNumber": {
					"type": "string"
				},
				"customerId": {
					"type": "string"
				},
				"accountName": {
					"type": "string"
				},
				"currency": {
					"type": "string"
				},
				"accountOpeningDate": {
					"type": "string"
				},
				"lastTransactionDate": {
					"type": "string"
				},
				"accountType": {
					"type": "string"
				},
				"bvn": {
					"type": "string"
				},
				"fullName": {
					"type": "string"
				},
				"phoneNumber": {
					"type": "string"
				},
				"email": {
					"type": "string"
				},
				"status": {
					"type": "string"
				},
				"expansionFields": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/ExpansionField"
					}
				}
			},
			"required": [
				"accountNumber",
				"customerId",
				"accountName",
				"currency",
				"accountOpeningDate",
				"lastTransactionDate",
				"accountType",
				"bvn",
				"fullName",
				"phoneNumber",
				"email",
				"status",
				"expansionFields"
			]
		},
		"GetSpecificAccount-Success": {
			"title": "GetSpecificAccount-Success",
			"example": {
				"status": "00",
				"message": "The process was completed successully",
				"data": [
					{
						"accountNumber": "1010119934",
						"customerId": "1234567890",
						"accountName": "Ado John Sule",
						"currency": "NGN",
						"accountOpeningDate": "2017-01-13",
						"lastTransactionDate": "2019-03-15T07:05:59.524Z",
						"accountType": "Wonderful Savings Account",
						"bvn": "221422456789",
						"fullName": "Ado John Sule",
						"phoneNumber": 342391,
						"email": "ado.john@example.com",
						"status": "Active",
						"expansionFields": [
							{
								"id": "Some random ID",
								"description": "Some random text",
								"type": "Some random type",
								"value": "Some value"
							}
						]
					}
				]
			},
			"type": "object",
			"properties": {
				"status": {
					"type": "string"
				},
				"message": {
					"type": "string"
				},
				"data": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/Data19"
					}
				}
			},
			"required": [
				"status",
				"message",
				"data"
			]
		},
...
	},
...
}